Fix incorrect logic in plan dependency recording
Both 7598f7a68 and 55bae7fe8 mistakenly tried to record a plan dependency
on a function but mistakenly inverted the OidIsValid test.  This meant
that we'd record a dependency only when the function's Oid was
InvalidOid.  Clearly this was meant to *not* record the dependency in
that case.

7598f7a68 made this mistake first, then in v15 55bae7fe8 copied the same
mistake.

Reported-by: Tom Lane
Backpatch-through: 14, where 7598f7a68 first made this mistake
Discussion: https://postgr.es/m/2277537.1694301772@sss.pgh.pa.us